Add text for entropy coding and OBMC.
Also includes additional edits.
Because it's really hard to talk about MC without talking about blocks, I
removed the text about, "The definition of block is a square transform block."
As long as each section specifies "transform block" or "MC block" the first
few times it mentions blocks, this should not be too difficult to follow (the
alternative being to start calling them TUs and PUs).